requirements.txt
appdirs==1.4.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
distlib==0.3.0
filelock==3.0.12
Flask==1.1.1
gunicorn==20.0.4
hurry.filesize==0.9
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nltk==3.4.5
pafy==0.5.5
pipenv==2018.11.26
pylint==2.4.4
requests==2.22.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.3
Werkzeug==1.0.0
wrapt==1.11.2
youtube-dl==2020.3.8
zipp==3.1.0
